requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
colorama==0.4.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.1
factory-boy==3.2.1
Faker==15.2.0
inflection==0.5.1
iniconfig==1.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.2
